# Optimize base64/hex decoding by pre-allocating output buffers (~2x faster)