## Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and enhancements and this helps us generate change logs for our releases. You can link an issue to this PR using the GitHub syntax. For example `Closes #123` indicates that this PR will close issue #123.
-->

Closes #.

## Rationale for this change

<!--
 Why are you proposing this change? If this is already explained clearly in the issue then this section is not needed.
 Explaining clearly why changes are proposed helps reviewers understand your changes and offer better suggestions for fixes.  
-->

It is generally faster to make a big allocation up front, rather than making many small allocations.

## What changes are included in this PR?

<!--
There is no need to duplicate the description in the issue here but it is sometimes worth providing a summary of the individual changes in this PR.
-->

- Add a benchmark
- Refactor to reduce code duplication
- Change decoding methods to write into pre-allocated buffer

## Are these changes tested?

<!--
We typically require tests for all PRs in order to:
1. Prevent the code from being accidentally broken by subsequent changes
2. Serve as another way to document the expected behavior of the code

If tests are not included in your PR, please explain why (for example, are they covered by existing tests)?
-->
Relying on existing SQL tests.

## Are there any user-facing changes?

Yes, base64 and hex decoding is ~2x faster:
```
base64_decode/1024      time:   [26.362 µs 26.459 µs 26.597 µs]
                        change: [-48.501% -47.320% -45.897%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  4 (4.00%) high mild
  3 (3.00%) high severe

hex_decode/1024         time:   [92.789 µs 92.904 µs 93.035 µs]
                        change: [-58.973% -58.895% -58.816%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  3 (3.00%) high mild
  2 (2.00%) high severe

base64_decode/4096      time:   [100.23 µs 100.30 µs 100.38 µs]
                        change: [-62.678% -62.609% -62.545%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) low severe
  1 (1.00%) high mild
  1 (1.00%) high severe

hex_decode/4096         time:   [373.11 µs 377.63 µs 383.37 µs]
                        change: [-58.265% -58.081% -57.870%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 12 outliers among 100 measurements (12.00%)
  1 (1.00%) low severe
  5 (5.00%) high mild
  6 (6.00%) high severe

base64_decode/8192      time:   [205.00 µs 205.26 µs 205.62 µs]
                        change: [-62.173% -61.549% -60.938%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  2 (2.00%) high mild
  5 (5.00%) high severe

hex_decode/8192         time:   [735.37 µs 737.54 µs 740.51 µs]
                        change: [-59.434% -59.332% -59.214%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  2 (2.00%) high mild
  5 (5.00%) high severe
```